Removed float to have an better approch with the info from the countries_data.py

+++ Country_data.py
@@ -3,12 +3,12 @@ from Countries_info import countries
 
 class Country:
     def __init__(self, name, population, capital, area):
-        if int(population) < 0 or float(area) < 0:
+        if int(population) < 0 or int(area) < 0:
             raise ValueError("Population and area must be non-negative")
         self.name = name
         self.population = int(population) # use int to handle numbers
         self.capital = capital
-        self.area = float(area)
+        self.area = int(area)  # Can use float if your data has float numbers
         self.subdivisions = []  # A list to store subdivisions
 
     def __str__(self):  # Returns a string representation of the Country
@@ -45,12 +45,12 @@ class Country:
 
 class Subdivision:
     def __init__(self, name, population, capital, area):
-        if int(population) < 0 or float(area) < 0:
+        if int(population) < 0 or int(area) < 0:
             raise ValueError("Population and area must be non-negative")
         self.name = name
         self.population = int(population) # use int to handle numbers
         self.capital = capital
-        self.area = float(area)
+        self.area = int(area) # Can use float if your data has float numbers
         self.subdivisions = []  # This stores the smaller subdivisions
 
     def __str__(self):  # Returns a string representation of the subdivisions
